@@ -1,12 +1,13 @@
 {
   "name": "serverless-offline-sns",
-  "version": "0.76.0",
+  "version": "0.77.0",
   "description": "Serverless plugin to run a local SNS server and call lambdas with events notifications.",
-  "main": "dist/index.js",
+  "exports": "./dist/index.js",
+  "type": "module",
   "scripts": {
     "build": "tsc -p src",
     "watch": "tsc -p src -w",
-    "test": "nyc ts-mocha \"test/**/*.ts\" --slow 300 -p src/tsconfig.json",
+    "test": "nyc ts-mocha -r ts-node/register --loader=ts-node/esm --no-warnings=ExperimentalWarning --experimental-specifier-resolution=node \"test/**/*.ts\" -p test/tsconfig.json",
     "lint": "tslint -c tslint.json 'src/**/*'",
     "prepare": "yarn run lint && yarn test && yarn build",
     "prettier": "npx prettier --write src test",
@@ -33,39 +34,45 @@
   },
   "homepage": "https://github.com/mj1618/serverless-offline-sns#readme",
   "dependencies": {
-    "aws-sdk": "^2.943.0",
-    "body-parser": "^1.19.0",
+    "aws-sdk": "^2.1476.0",
+    "body-parser": "^1.20.2",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "lodash": "^4.17.21",
-    "node-fetch": "^2.6.1",
-    "serverless": "^2.51.2",
-    "shelljs": "^0.8.4",
-    "uuid": "^8.3.2",
+    "node-fetch": "^3.3.2",
+    "serverless": "^3.35.2",
+    "shelljs": "^0.8.5",
+    "uuid": "^9.0.1",
     "xml": "^1.0.1"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.21",
-    "@types/express": "^4.17.13",
-    "@types/mocha": "^8.2.3",
-    "@types/node": "^16.3.1",
-    "@types/shelljs": "^0.8.9",
-    "all-contributors-cli": "^6.20.0",
-    "aws-sdk-mock": "^5.2.1",
-    "chai": "^4.3.4",
-    "handlebars": "4.7.7",
+    "@types/chai": "^4.3.9",
+    "@types/cors": "^2.8.15",
+    "@types/express": "^4.17.20",
+    "@types/mocha": "^10.0.3",
+    "@types/node": "^20.8.7",
+    "@types/node-fetch": "^2.6.7",
+    "@types/serverless": "^3.12.17",
+    "@types/shelljs": "^0.8.14",
+    "@types/sinon": "^10.0.20",
+    "@types/uuid": "^9.0.6",
+    "@types/xml": "^1.0.10",
+    "all-contributors-cli": "^6.26.1",
+    "aws-sdk-mock": "^5.8.0",
+    "chai": "^4.3.10",
+    "handlebars": "4.7.8",
     "istanbul": "^0.4.5",
-    "mocha": "^9.0.2",
+    "mocha": "^10.2.0",
     "nyc": "^15.1.0",
-    "prettier": "2.3.2",
-    "sinon": "^11.1.1",
-    "source-map-support": "^0.5.19",
-    "ts-loader": "^9.2.3",
+    "prettier": "3.0.3",
+    "sinon": "^16.1.1",
+    "source-map-support": "^0.5.21",
+    "ts-loader": "^9.5.0",
     "ts-mocha": "^10.0.0",
-    "ts-node": "^10.1.0",
+    "ts-node": "^10.9.1",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^4.3.5"
+    "typescript": "^5.2.2"
   },
   "nyc": {
     "extension": [
