@@ -1,10 +1,10 @@
-const ServerlessOfflineSns = require("../../src/index");
-import * as AWS from "aws-sdk";
-import * as AWSMock from "aws-sdk-mock";
+import ServerlessOfflineSns from "../../src/index.js";
+import AWSMock from 'aws-sdk-mock';
 import { expect } from "chai";
-import { assert, spy } from "sinon";
-import * as handler from "../mock/handler";
-import * as state from "../mock/mock.state";
+import sinon from 'sinon';
+import * as handler from "../mock/handler.js";
+import * as state from "../mock/mock.state.js";
+import AWS from "aws-sdk";
 
 let plugin;
 
@@ -23,7 +23,6 @@ describe("test", () => {
 
   it("should start on offline start", async () => {
     plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
     });
     await plugin.hooks["before:offline:start:init"]();
     await plugin.hooks["after:offline:start:end"]();
@@ -31,17 +30,25 @@ describe("test", () => {
 
   it("should start on command start", async () => {
     plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
     });
     plugin.hooks["offline-sns:start:init"]();
     await new Promise((res) => setTimeout(res, 100));
     await plugin.hooks["offline-sns:start:end"]();
   });
 
-  it("should send event to topic ARN", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+  it('should send event to topic ARN', async () => {
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
+    const snsAdapter = await plugin.start();
+    await snsAdapter.publish(
+      `arn:aws:sns:us-east-1:${accountId}:test-topic`,
+      "{}"
+    );
+    await new Promise((res) => setTimeout(res, 500));
+    expect(state.getPongs()).to.eq(2);
+  });
+
+  it('should send event to topic ARN', async () => {
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
       `arn:aws:sns:us-east-1:${accountId}:test-topic`,
@@ -52,9 +59,7 @@ describe("test", () => {
   });
 
   it("should send event to target ARN", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     await snsAdapter.publishToTargetArn(
       `arn:aws:sns:us-east-1:${accountId}:test-topic`,
@@ -65,9 +70,7 @@ describe("test", () => {
   });
 
   it("should send event with pseudo parameters", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
       "arn:aws:sns:us-east-1:#{AWS::AccountId}:test-topic",
@@ -78,9 +81,7 @@ describe("test", () => {
   });
 
   it("should send event with MessageAttributes and subject", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
       `arn:aws:sns:us-east-1:${accountId}:test-topic`,
@@ -107,9 +108,7 @@ describe("test", () => {
   });
 
   it("should return a valid response to publish", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     const snsResponse = await snsAdapter.publish(
       `arn:aws:sns:us-east-1:${accountId}:test-topic`,
@@ -122,9 +121,7 @@ describe("test", () => {
   });
 
   it("should send a message to a E.164 phone number", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     const snsResponse = await snsAdapter.publishToPhoneNumber(
       `+10000000000`,
@@ -158,8 +155,7 @@ describe("test", () => {
 
   it("should use the custom host for subscription urls", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "pongHandler", "0.0.0.0"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "pongHandler", "0.0.0.0")
     );
     const snsAdapter = await plugin.start();
     const response = await snsAdapter.listSubscriptions();
@@ -171,8 +167,7 @@ describe("test", () => {
 
   it("should use the custom subscribe endpoint for subscription urls", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "pongHandler", "0.0.0.0", "anotherHost"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "pongHandler", "0.0.0.0", "anotherHost")
     );
     const snsAdapter = await plugin.start();
     const response = await snsAdapter.listSubscriptions();
@@ -183,9 +178,7 @@ describe("test", () => {
   });
 
   it("should unsubscribe", async () => {
-    plugin = new ServerlessOfflineSns(createServerless(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerless(accountId));
     const snsAdapter = await plugin.start();
     await plugin.unsubscribeAll();
     await snsAdapter.publish(
@@ -198,8 +191,7 @@ describe("test", () => {
 
   it("should read env variable", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "envHandler"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "envHandler")
     );
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
@@ -211,8 +203,7 @@ describe("test", () => {
 
   it("should read env variable for function", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "envHandler"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "envHandler")
     );
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
@@ -224,8 +215,7 @@ describe("test", () => {
 
   it("should convert pseudo param on load", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "pseudoHandler"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "pseudoHandler")
     );
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
@@ -237,32 +227,8 @@ describe("test", () => {
     );
   });
 
-  it("should completely reload the module every time if cache invalidation is enabled", async () => {
-    plugin = new ServerlessOfflineSns(
-      createServerlessCacheInvalidation(accountId),
-      { skipCacheInvalidation: [/mock\.state/] }
-    );
-
-    const snsAdapter = await plugin.start();
-    await snsAdapter.publish(
-      `arn:aws:sns:us-east-1:${accountId}:test-topic`,
-      "{}"
-    );
-    await new Promise((res) => setTimeout(res, 100));
-    expect(state.getPongs()).to.eq(1, "wrong number of pongs (first check)");
-
-    await snsAdapter.publish(
-      `arn:aws:sns:us-east-1:${accountId}:test-topic`,
-      "{}"
-    );
-    await new Promise((res) => setTimeout(res, 100));
-    expect(state.getPongs(), "wrong number of pongs (second check)").to.eq(1);
-  });
-
   it("should send event to handlers with more than one dot in the filename", async () => {
-    plugin = new ServerlessOfflineSns(createServerlessMultiDot(accountId), {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(createServerlessMultiDot(accountId));
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
       `arn:aws:sns:us-east-1:${accountId}:multi-dot-topic`,
@@ -271,17 +237,17 @@ describe("test", () => {
     await new Promise((res) => setTimeout(res, 100));
     expect(state.getPongs()).to.eq(1);
   });
-
-  it("should support async handlers with no callback", async () => {
-    plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "asyncHandler"),
-      { skipCacheInvalidation: true }
-    );
+  
+  it('should support async handlers with no callback', async () => {
+    plugin = new ServerlessOfflineSns(createServerless(accountId, "asyncHandler"), {
+      skipCacheInvalidation: true,
+    });
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
       `arn:aws:sns:us-east-1:${accountId}:test-topic-async`,
       "{}"
     );
+    await new Promise((res) => setTimeout(res, 100));
     expect(await state.getResult()).to.eq(
       `arn:aws:sns:us-east-1:${accountId}:test-topic-async`
     );
@@ -308,8 +274,7 @@ describe("test", () => {
 
   it("should send event when filter policies exist and pass", async () => {
     plugin = new ServerlessOfflineSns(
-      createServerlessWithFilterPolicies(accountId),
-      { skipCacheInvalidation: true }
+      createServerlessWithFilterPolicies(accountId)
     );
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
@@ -360,15 +325,14 @@ describe("test", () => {
     );
     await new Promise((res) => setTimeout(res, 100));
     expect(state.getPongs()).to.eq(0);
+
   });
 
   it("should not wrap the event when the sub's raw message delivery is true", async () => {
     const serverless = createServerless(accountId);
     serverless.service.functions.pong4.events[0].sns["rawMessageDelivery"] =
       "true";
-    plugin = new ServerlessOfflineSns(serverless, {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(serverless);
 
     const snsAdapter = await plugin.start();
     await snsAdapter.publish(
@@ -392,12 +356,11 @@ describe("test", () => {
   });
 
   it("should subscribe", async () => {
-    const spySendMessage = spy();
+    const spySendMessage = sinon.spy();
     AWSMock.setSDKInstance(AWS);
     AWSMock.mock("SQS", "sendMessage", spySendMessage);
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "envHandler"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "envHandler")
     );
     const snsAdapter = await plugin.start();
     await plugin.subscribeAll();
@@ -406,31 +369,29 @@ describe("test", () => {
       "{}"
     );
     await new Promise((res) => setTimeout(res, 100));
-    assert.calledOnce(spySendMessage);
-    assert.calledWith(spySendMessage, {
+    sinon.assert.calledOnce(spySendMessage);
+    sinon.assert.calledWith(spySendMessage, {
       QueueUrl: "http://127.0.0.1:4002/undefined",
       MessageBody: "{}",
       MessageAttributes: {},
     });
     AWSMock.restore("SQS", "sendMessage");
+
   });
 
   it("should handle empty resource definition", async () => {
     const serverless = createServerless(accountId);
     serverless.service.resources = undefined;
-    plugin = new ServerlessOfflineSns(serverless, {
-      skipCacheInvalidation: true,
-    });
+    plugin = new ServerlessOfflineSns(serverless);
     await plugin.start();
   });
 
   it("should handle messageGroupId", async () => {
-    const spySendMessage = spy();
+    const spySendMessage = sinon.spy();
     AWSMock.setSDKInstance(AWS);
     AWSMock.mock("SQS", "sendMessage", spySendMessage);
     plugin = new ServerlessOfflineSns(
-      createServerless(accountId, "envHandler"),
-      { skipCacheInvalidation: true }
+      createServerless(accountId, "envHandler")
     );
     const snsAdapter = await plugin.start();
     await plugin.subscribeAll();
@@ -443,8 +404,8 @@ describe("test", () => {
       "messageGroupId-here"
     );
     await new Promise((res) => setTimeout(res, 100));
-    assert.calledOnce(spySendMessage);
-    assert.calledWith(spySendMessage, {
+    sinon.assert.calledOnce(spySendMessage);
+    sinon.assert.calledWith(spySendMessage, {
       QueueUrl: "http://127.0.0.1:4002/undefined",
       MessageBody: "{}",
       MessageAttributes: {},
@@ -457,21 +418,18 @@ describe("test", () => {
 const createServerless = (
   accountId: number,
   handlerName: string = "pongHandler",
-  host: string|null = null,
-  subscribeEndpoint: string|null = null
+  host: string | null = null,
+  subscribeEndpoint: string | null = null
 ) => {
   return {
-    config: {
-      // Since this is not really being used, it should be removed
-      skipCacheInvalidation: true,
-    },
+    config: {},
     service: {
       custom: {
         "serverless-offline-sns": {
           debug: true,
           port: 4002,
-          accountId: accountId,
-          host: host,
+          accountId,
+          host,
           "sns-subscribe-endpoint": subscribeEndpoint,
         },
       },
@@ -592,12 +550,10 @@ const createServerless = (
 const createServerlessCacheInvalidation = (
   accountId: number,
   handlerName: string = "pongHandler",
-  host: string = null
+  host = null
 ) => {
   return {
-    config: {
-      skipCacheInvalidation: [/mock\.state/],
-    },
+    config: {},
     service: {
       custom: {
         "serverless-offline-sns": {
@@ -639,12 +595,10 @@ const createServerlessCacheInvalidation = (
 const createServerlessMultiDot = (
   accountId: number,
   handlerName: string = "pongHandler",
-  host: string = null
+  host = null
 ) => {
   return {
-    config: {
-      skipCacheInvalidation: true,
-    },
+    config: {},
     service: {
       custom: {
         "serverless-offline-sns": {
@@ -684,9 +638,7 @@ const createServerlessMultiDot = (
 
 const createServerlessBad = (accountId: number) => {
   return {
-    config: {
-      skipCacheInvalidation: true,
-    },
+    config: {},
     service: {
       custom: {
         "serverless-offline-sns": {
@@ -725,20 +677,18 @@ const createServerlessBad = (accountId: number) => {
 const createServerlessWithFilterPolicies = (
   accountId: number,
   handlerName: string = "pongHandler",
-  host: string = null,
+  host = null,
   subscribeEndpoint = null
 ) => {
   return {
-    config: {
-      skipCacheInvalidation: true,
-    },
+    config: {},
     service: {
       custom: {
         "serverless-offline-sns": {
           debug: true,
           port: 4002,
-          accountId: accountId,
-          host: host,
+          accountId,
+          host,
           "sns-subscribe-endpoint": subscribeEndpoint,
         },
       },
