@@ -1,11 +1,11 @@
-import * as AWS from "aws-sdk";
+import AWS from "aws-sdk";
 import {
   ListSubscriptionsResponse, ListTopicsResponse, MessageAttributeMap
-} from "aws-sdk/clients/sns.d";
-import * as _ from "lodash";
+} from "aws-sdk/clients/sns.d.js";
+import _ from "lodash";
 import fetch from "node-fetch";
-import { createMessageId, createSnsLambdaEvent } from "./helpers";
-import { IDebug, ISNSAdapter } from "./types";
+import { createMessageId, createSnsLambdaEvent } from "./helpers.js";
+import { IDebug, ISNSAdapter } from "./types.js";
 
 export class SNSAdapter implements ISNSAdapter {
   private sns: AWS.SNS;
@@ -156,8 +156,7 @@ export class SNSAdapter implements ISNSAdapter {
       if (req.body.SubscribeURL) {
         this.debug("Visiting subscribe url: " + req.body.SubscribeURL);
         return fetch(req.body.SubscribeURL, {
-          method: "GET",
-          timeout: 0,
+          method: "GET"
         }).then((fetchResponse) => this.debug("Subscribed: " + fetchResponse));
       }
 
@@ -171,7 +170,7 @@ export class SNSAdapter implements ISNSAdapter {
           this.sent(response);
         }
       };
-      const maybePromise = getHandler()(
+      const maybePromise = getHandler(
         event,
         this.createLambdaContext(fn, sendIt),
         sendIt
